requirements.txt
requests==2.31.0
python-dotenv==1.0.1
rich==13.7.1
SpeechRecognition==3.10.3
PyAudio==0.2.14
gTTS==2.5.1
playsound==1.2.2
# Optional extras:
# vosk==0.3.45
# faster-whisper==1.0.3
# pygame==2.6.0
gradio==4.44.0
